Renaming Domain/Models to Domain/Entities.
