+++ app/Modules/Character/Domain/Entities/ValueObjects/Name.php
@@ -1,7 +1,7 @@
 <?php
 
 
-namespace App\Modules\Character\Domain\Models\ValueObjects;
+namespace App\Modules\Character\Domain\Entities\ValueObjects;
 
 
 class Name
