+++ app/Modules/Character/Infrastructure/Repositories/LocationRepository.php
@@ -3,7 +3,7 @@
 namespace App\Modules\Character\Infrastructure\Repositories;
 
 use App\Modules\Character\Domain\Contracts\LocationRepositoryInterface;
-use App\Modules\Character\Domain\Models\Location;
+use App\Modules\Character\Domain\Entities\Location;
 
 class LocationRepository implements LocationRepositoryInterface
 {
