+++ app/Modules/User/Domain/Contracts/UserRepositoryInterface.php
@@ -3,7 +3,7 @@
 namespace App\Modules\User\Domain\Contracts;
 
 
-use App\Modules\User\Domain\Models\User;
+use App\Modules\User\Domain\Entities\User;
 
 interface UserRepositoryInterface
 {
