+++ app/Modules/User/Domain/Factories/UserFactory.php
@@ -4,7 +4,7 @@
 namespace App\Modules\User\Domain\Factories;
 
 use App\Factories\GeneratesUuid;
-use App\Modules\User\Domain\Models\User;
+use App\Modules\User\Domain\Entities\User;
 use App\Modules\User\Domain\Requests\CreateUserRequest;
 
 class UserFactory
